API/requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
Django==5.0.4
psycopg==3.1.19
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
